requirements.txt
flask
dotenv
google-generativeai 
gunicorn